+++ packages/hooks/src/index.ts
@@ -1,1 +1,3 @@
 export { default as useSearchAndTable } from './useSearchAndTable';
+export { default as useDownload } from './useDownload';
+export { handleDownload } from './useDownload';