+++ packages/hooks/src/useDownload/index.ts
@@ -1,0 +1,5 @@
+import useDownload from './useDownload';
+import handleDownload from './handleDownload';
+
+export { handleDownload };
+export default useDownload;